refactor: Split database.rs into separate modules
